Add ResultPage typescript migrations
